change: Move Markwon core.markwon
This confines the non-incremental parts of the build that use kapt to a
single module.
