+++ core/markwon/src/main/kotlin/app/pachli/core/markwon/MarkwonFactory.kt
@@ -1,0 +1,218 @@
+/*
+ * Copyright (c) 2026 Pachli Association
+ *
+ * This file is a part of Pachli.
+ *
+ * This program is free software; you can redistribute it and/or modify it under the terms of the
+ * GNU General Public License as published by the Free Software Foundation; either version 3 of the
+ * License, or (at your option) any later version.
+ *
+ * Pachli is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even
+ * the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General
+ * Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License along with Pachli; if not,
+ * see <http://www.gnu.org/licenses>.
+ */
+
+package app.pachli.core.markwon
+
+import android.content.Context
+import android.graphics.Color
+import androidx.core.text.method.LinkMovementMethodCompat
+import app.pachli.core.markwon.PreProcessMastodonHtml.processMarkdown
+import com.google.android.material.color.MaterialColors
+import com.mohamedrejeb.ksoup.entities.KsoupEntities
+import io.noties.markwon.AbstractMarkwonPlugin
+import io.noties.markwon.Markwon
+import io.noties.markwon.MarkwonConfiguration
+import io.noties.markwon.MarkwonPlugin
+import io.noties.markwon.RenderProps
+import io.noties.markwon.SoftBreakAddsNewLinePlugin
+import io.noties.markwon.core.MarkwonTheme
+import io.noties.markwon.core.spans.CodeBlockSpan
+import io.noties.markwon.core.spans.CodeSpan
+import io.noties.markwon.ext.latex.JLatexMathPlugin
+import io.noties.markwon.ext.strikethrough.StrikethroughPlugin
+import io.noties.markwon.html.HtmlPlugin
+import io.noties.markwon.html.HtmlTag
+import io.noties.markwon.html.tag.SimpleTagHandler
+import io.noties.markwon.inlineparser.MarkwonInlineParserPlugin
+import io.noties.markwon.movement.MovementMethodPlugin
+import io.noties.markwon.syntax.Prism4jThemeDefault
+import io.noties.markwon.syntax.SyntaxHighlightPlugin
+import io.noties.prism4j.Prism4j
+import io.noties.prism4j.annotations.PrismBundle
+import java.util.Collections
+import kotlin.math.roundToInt
+
+@PrismBundle(includeAll = true, grammarLocatorClassName = ".MySuperGrammerLocator")
+object MarkwonFactory {
+    fun makeMarkwon(context: Context, textSize: Float) = Markwon.builder(context)
+        .usePlugin(MovementMethodPlugin.create(LinkMovementMethodCompat.getInstance()))
+        .usePlugin(HtmlPlugin.create())
+        .usePlugin(
+            object : AbstractMarkwonPlugin() {
+                override fun configure(registry: MarkwonPlugin.Registry) {
+                    registry.require(HtmlPlugin::class.java) { htmlPlugin ->
+                        htmlPlugin.addHandler(CodeTagHandler)
+                        htmlPlugin.addHandler(PreTagHandler)
+                    }
+                }
+            },
+        )
+        .usePlugin(SoftBreakAddsNewLinePlugin.create())
+        .usePlugin(MarkwonInlineParserPlugin.create())
+        .usePlugin(
+            JLatexMathPlugin.create(textSize) {
+                it.inlinesEnabled(true)
+            },
+        )
+        .usePlugin(
+            SyntaxHighlightPlugin.create(
+                Prism4j(MySuperGrammerLocator()),
+                Prism4jThemeDefault.create(),
+            ),
+        )
+        .usePlugin(StrikethroughPlugin.create())
+        .usePlugin(PreProcessMastodonHtml)
+        // Has to be at the end of the list, see https://github.com/noties/Markwon/issues/494
+        .usePlugin(PachliMarkwonTheme(context, textSize))
+        .build()
+}
+
+private object CodeTagHandler : SimpleTagHandler() {
+    override fun getSpans(configuration: MarkwonConfiguration, renderProps: RenderProps, tag: HtmlTag): Any {
+        return CodeSpan(configuration.theme())
+    }
+
+    override fun supportedTags(): Collection<String> = Collections.singleton("code")
+}
+
+private object PreTagHandler : SimpleTagHandler() {
+    override fun getSpans(configuration: MarkwonConfiguration, renderProps: RenderProps, tag: HtmlTag): Any {
+        return CodeBlockSpan(configuration.theme())
+    }
+
+    override fun supportedTags(): Collection<String> = Collections.singleton("pre")
+}
+
+/**
+ * Markwon theme plugin to set foreground and background colours for code blocks
+ * (inline and fenced) from the app theme, to ensure code blocks are still legible
+ * in light and dark themes.
+ */
+private class PachliMarkwonTheme(private val context: Context, private val textSize: Float) : AbstractMarkwonPlugin() {
+    override fun configureTheme(builder: MarkwonTheme.Builder) {
+        val blockQuoteStripeColor = MaterialColors.getColor(
+            context,
+            androidx.appcompat.R.attr.colorPrimary,
+            Color.BLACK,
+        )
+
+        val codeBackgroundColor = MaterialColors.getColor(
+            context,
+            com.google.android.material.R.attr.colorSurfaceContainerLow,
+            Color.WHITE,
+        )
+
+        val codeTextColor = MaterialColors.getColor(
+            context,
+            com.google.android.material.R.attr.colorOnSurfaceVariant,
+            Color.BLACK,
+        )
+
+        builder
+            .blockMargin(textSize.roundToInt())
+            .blockQuoteColor(blockQuoteStripeColor)
+            .blockQuoteWidth((textSize / 4).roundToInt())
+            .codeBackgroundColor(codeBackgroundColor)
+            .codeTextColor(codeTextColor)
+            .codeBlockBackgroundColor(codeBackgroundColor)
+            .codeBlockTextColor(codeTextColor)
+    }
+}
+
+/**
+ * Markwon plugin that pre-processes HTML from Mastodon before processing as Markdown.
+ *
+ * @see [processMarkdown]
+ */
+private object PreProcessMastodonHtml : AbstractMarkwonPlugin() {
+    /** Match the contents of a fenced code block. */
+    private val rxCodeBlock =
+        """^```(.*?)^```""".toRegex(setOf(RegexOption.DOT_MATCHES_ALL, RegexOption.MULTILINE))
+
+    /** Match any start or end tag. */
+    private val rxTag = """</?[^>]+?>""".toRegex(RegexOption.DOT_MATCHES_ALL)
+
+    /** Match any inline code element (i.e., between two literal backquotes. */
+    private val rxLiteral = """(?<!`)`(.+?)`""".toRegex(setOf(RegexOption.MULTILINE))
+
+    /**
+     * Match any opening `p` element, with optional whitespace before it, and optional
+     * whitespace immediately after the opening tag.
+     *
+     * The optional whitespace after the tag is to catch `<p>   ...`, which would
+     * appear as a code block because of the indentation.
+     */
+    private val rxPEl = """\s*<p\s*>\s*""".toRegex()
+
+    /** Match `<br>`, `<br/>` (with optional spaces). */
+    private val rxBrEl = """<br\s*/?>""".toRegex()
+
+    /**
+     * Match `~~~` at the start of a line. These are an alternative mechanism for
+     * starting a fenced code block. They seem to be much less frequent than
+     * using backticks.
+     */
+    private val rxThreeTilde = """^~~~""".toRegex(RegexOption.MULTILINE)
+
+    /**
+     * Processes [input] and removes/replaces some HTML content.
+     *
+     * Any Markdown in [input] will be wrapped in HTML block elements. Undo those,
+     * so the content becomes "Markdown that possibly contains embedded HTML" and not
+     * "HTML that might contain some Markdown directives".
+     */
+    override fun processMarkdown(input: String): String {
+        // Known differences from Mastodon web rendering
+        //
+        // - Links in fenced code blocks are not clickable
+        //   Eg., a hashtag in a fenced code block is not clickable
+        val processed = input
+            // Remove <p> with any preceeding whitespace (just in case a paragraph was
+            // indented -- not removing the whitespace could cause it to be treated as
+            // a code block).
+            .replace(rxPEl, "")
+            // </p> ends a paragraph, Markdown needs two blank lines as separators.
+            .replace("</p>", "\n\n")
+            // <br> and variations become "\n".
+            .replace(rxBrEl, "\n")
+            // Convert leading quote markers from entities to ">".
+            .replace("&gt; ", "> ")
+            // Three ~ at the start of a line are unlikely to be a fenced code block
+            // and are more likely to be decoration. Escape them so they are not parsed
+            // as code.
+            //
+            // https://dair-community.social/@emilymbender/114172441506624981
+            .replace(rxThreeTilde, Regex.escapeReplacement("""\~\~\~"""))
+            // Hack for Mathstodon. Mathstodon uses `\[...\]` for block latex content
+            // and `\(...\)` for inline latex content (those are literal backslash and
+            // bracket/brace characters). Rewrite those to block or inline level `$$`
+            // strings so the JLatexMath plugin will parse them.
+            .replace("""\[""", "\n\n$$\n")
+            .replace("""\]""", "\n$$\n\n")
+            .replace("""\(""", "$$")
+            .replace("""\)""", "$$")
+            // HTML in fenced code blocks is treated literally by Markwon.
+            // So remove all HTML tags inside fenced blocks (keep the content).
+            //
+            // There may be HTML entities that need to be converted.
+            .replace(rxCodeBlock) { KsoupEntities.decodeHtml(it.value.replace(rxTag, "")) }
+            // Convert any HTML entities in inline code.
+            .replace(rxLiteral) { KsoupEntities.decodeHtml(it.value) }
+
+        return processed
+    }
+}
