+++ settings.gradle.kts
@@ -57,6 +57,7 @@ include(":core:database")
 include(":core:designsystem")
 include(":core:domain")
 include(":core:eventhub")
+include(":core:markwon")
 include(":core:model")
 include(":core:preferences")
 include(":core:navigation")
